requirements.txt
#Backend Server
Flask
Flask-Cors

#LangChain Core and Integrations
langchain
langchain-openai
langchain-community
sqlalchemy

#LangChain Third-Party Integrations
langchain-google-community

#Web Scraping & HTTP
beautifulsoup4
requests
httpx==0.27.0

#Environment & Utilities
python-dotenv

#Google API for Search
google-api-python-client

#OpenAI library (pinned for compatibility)
openai<2.0.0,>=1.12.0